genericize vuln report handling, split osv into a subpackage
